+++ tests/assertions/assert_text_contains.test.js
@@ -86,4 +86,46 @@ describe("Assert Text Contains", function() {
             await browser.assert.not.textContains(".container p", "first", "not textcontains fails");
         }, `[assert.not.textContains] not textcontains fails`);
     });
+
+    it("Multiple Texts in Array With Regex", async() => {
+        await browser.assert.textContains("p", ["My first", "My second"]);
+        await browser.assert.textContains("p", ["My first"]);
+    });
+
+    it("Multiple Texts in Array Throws", async() => {
+        await utils.assertThrowsAssertionAsync(async() => {
+            await browser.assert.textContains("p", ["My first paragraph", "My pa2ragraph"]);
+        }, `[assert.textContains] Expected element "p" to contain text "My pa2ragraph", "My first paragraph My second paragraph" found.`);
+    });
+
+    it("Not Multiple Texts in Array", async() => {
+        await browser.assert.not.textContains("p", ["not a correct text", "another incorrect text"]);
+    });
+
+    it("Not Multiple Texts in Array Throws", async() => {
+        await utils.assertThrowsAssertionAsync(async() => {
+            await browser.assert.not.textContains("p", ["incorrect", "My first", "not a correct text"]);
+        }, `[assert.not.textContains] Expected element "p" to not contain text "My first".`);
+    });
+
+    it("Text With Empty Parameter Array Throws", async() => {
+        await utils.assertThrowsAsync(async() => {
+            await browser.assert.textContains("h1", []);
+        }, `Error: [assert.textContains] Missing expected text for assertion.`);
+    });
+
+    it("Not Text With Empty Parameter Array Throws", async() => {
+        await utils.assertThrowsAsync(async() => {
+            await browser.assert.not.textContains("h1", []);
+        }, `Error: [assert.not.textContains] Missing expected text for assertion.`);
+    });
+
+    it("Text Contains Throws", async() => {
+        await utils.assertThrowsAssertionAsync(async() => {
+            await browser.assert.textContains(".container p", "My second");
+        }, `[assert.textContains] Expected element ".container p" to contain text "My second", "My first paragraph" found.`);
+        await utils.assertThrowsAssertionAsync(async() => {
+            await browser.assert.textContains("p", "My paragraph");
+        }, `[assert.textContains] Expected element "p" to contain text "My paragraph", "My first paragraph My second paragraph" found.`);
+    });
 });